+++ src/products/Laptop.java
@@ -7,14 +7,14 @@ public class Laptop extends ConsumerElectronics {
     private String cpuModel;
     private String os;
 
-    public Laptop(String id, String brand, String model, Double price, Double height, Double width, Double depth, int weight, Double reviewPoint, Double screenSize, int storageSize, int ramSize, String cpuModel, String os) {
+    public Laptop(String id, String brand, String model, Double price, Double height, Double width, Double depth, Double weight, Double reviewPoint, Double screenSize, int storageSize, int ramSize, String cpuModel, String os) {
         super(id, brand, model, price,"Laptop", height, width, depth, weight, reviewPoint,  screenSize ,storageSize);
         this.ramSize = ramSize;
         this.cpuModel = cpuModel;
         this.os = os;
     }
 
-    public Laptop(String brand, String model, Double price, Double height, Double width, Double depth, int weight, Double screenSize, int storageSize, int ramSize, String cpuModel, String os) {
+    public Laptop(String brand, String model, Double price, Double height, Double width, Double depth, Double weight, Double screenSize, int storageSize, int ramSize, String cpuModel, String os) {
         super(brand, model, price,"Laptop", height, width, depth, weight, screenSize ,storageSize);
         this.ramSize = ramSize;
         this.cpuModel = cpuModel;
