+++ src/products/MobilePhone.java
@@ -8,14 +8,14 @@ public class MobilePhone extends ConsumerElectronics {
     private String os;
     private int ramSize;
 
-    public MobilePhone(String id, String brand, String model, Double price, Double height, Double width, Double depth, int weight, Double reviewPoint, Double screenSize, int storageSize, int cameraResolution, String os, int ramSize) {
+    public MobilePhone(String id, String brand, String model, Double price, Double height, Double width, Double depth, Double weight, Double reviewPoint, Double screenSize, int storageSize, int cameraResolution, String os, int ramSize) {
         super(id, brand, model, price, "Mobile Phone", height, width, depth, weight, reviewPoint, screenSize, storageSize);
         this.cameraResolution = cameraResolution;
         this.os = os;
         this.ramSize = ramSize;
     }
 
-    public MobilePhone(String brand, String model, Double price, Double height, Double width, Double depth, int weight, Double screenSize, int storageSize, int cameraResolution, String os, int ramSize) {
+    public MobilePhone(String brand, String model, Double price, Double height, Double width, Double depth, Double weight, Double screenSize, int storageSize, int cameraResolution, String os, int ramSize) {
         super(brand, model, price, "Mobile Phone", height, width, depth, weight, screenSize, storageSize);
         this.cameraResolution = cameraResolution;
         this.os = os;
